NOT WORKING - commiting for reference
HomeAssistant's YAML implementation is broken, so it's not actually possible to use a template in this way.

If you configure

```
interval: 60
dns_container: hassio_dns
dns_template: |
.:53 {
    log {
        class error
    }
    errors
    loop

    hosts /config/hosts {
        fallthrough
    }
    template ANY AAAA local.hass.io hassio {
        rcode NOERROR
    }
    mdns
    forward . dns://192.168.1.253 {
        except local.hass.io
        policy sequential
    }
    cache 600
 }

.:5553 {
    log {
        class error
    }
    errors

    forward . dns://192.168.1.253  {
        except local.hass.io
    }
    cache 600
}
```

It'll convert to

```
dns_template: >-
  .:53 { log { class error } errors loop
```

The difference is that `>-` replaces newlines with spaces, so the config becomes invalid and coredns will crash out.

I'm going to swap the config option to a boolean - if active we'll copy from a template in config